+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 > **LoginRadius Ionic-2 SDK Change Log** provides information regarding what has changed, more specifically what changes, improvements and bug fix has been made to the SDK. For more details please refer to the [LoginRadius API Documention](https://docs.loginradius.com/api/v2/mobile-libraries/ionic-2)
 
 
+### Version 2.1.0
+Released on **November 30,  2018**
+
+##### Enhancements
+
+  - Added New Check Email,UserName and Phone Availability APIs.
+  - Added New Resend Email Verification API.
+  - Added New Verify OTP By Phone API.
+  - Added New Resend OTP By Phone API.
+  - Significantly improved code performance.
+  
+##### Breaking Changes
+For developers migrating from v2.0.0, there will be some breaking changes in terms of SDK implementation. In this version, we have updated endpoints and renamed "Auto Login" to "Smart Login", "No Registration/Simplified Registration" to "One touch Login" and "Instant Link Login" to "PasswordLess Login". Also, changed the methods of the above APIs accordingly.
+
 
 ### Version 2.0.0
-Released on **December 22,  2017**
+Released on **November 02,  2017**
 
 ##### Enhancements